Allow build-type Configure to work with Components
When components were introduced, cabal had support for
build-type: Configure, which allows to run a `configure`
script prior to building the package.  Upon the introduction
of components, it became obvious that we can't just
configure each and every component, as this easily runs
into the situation where the `configure` script is run
potentially concurrently for all components.

However, build-type is a global option for cabal files, and
is usually used to produce a <pkg>.buildinfo file.  This file
is then used to augment the Main Library only.

This change now tracks whether or not we are building a
component to the setup invocation, and if we do, ignores
a build-type: Configure built-type from the package
description.  For end users who want to depend on
package configured values, they will need to set their
sub libraries, executables, ... and other components to
depend on the main library, and import any configured
values from there.

For lib:Cabal, which doesn't know about components
when configure is invoked, we will continue to execute
configure, even though cabal will never use that info.